refactor: decompose oversized financial-overview and reports-page
financial-overview.tsx (356 lines) and reports-page.tsx (344) mixed
derivation, layout, and repeated blocks. Extract pure helpers into
financial-overview.utils (unit-tested), the repeated bar chart into
ComparisonBarChart, and the stat cards into StatCard; split the three
report tabs into their own files. No behavior change. Both files now
under 250 lines (198 and 65).
